# Avoid dynamic parse method dispatch for faster access