--- a/benchmark/parse_strip.yaml
+++ b/benchmark/parse_strip.yaml
@@ -1,5 +1,7 @@
 loop_count: 100
 contexts:
+  - gems:
+      csv: 3.3.0
   - name: "master"
     prelude: |
       $LOAD_PATH.unshift(File.expand_path("lib"))