--- a/benchmark/read.yaml
+++ b/benchmark/read.yaml
@@ -4,6 +4,8 @@ contexts:
       csv: 3.0.1
   - gems:
       csv: 3.0.2
+  - gems:
+      csv: 3.3.0
   - name: "master"
     prelude: |
       $LOAD_PATH.unshift(File.expand_path("lib"))