@kou 2024-07-06
Could you benchmark the `__send__` approach too?
I think that `__send__` is faster than `method.call` but we should measure them.